Started to tidy up testing class, changed testing class's name to match brief stipulation.
